% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cate_test.R
\name{plot_cate_test}
\alias{plot_cate_test}
\title{Variable importance of Bayesian Additive Regression Trees}
\usage{
plot_cate_test(.model, confounders)
}
\arguments{
\item{.model}{a model produced by bartCause::bartc()}

\item{confounders}{a character list of column names which should be considered the confounders. Must match the column names used to original fit .model.}
}
\value{
a list containing variable importance plot & plots for each conditional effect
}
\description{
Fit single regression tree on bartc() icates to produce variable importance plot & conditional effects plots. TODO: describe what the plot is and how it should be used
}
\examples{
data(lalonde, package = 'arm')
confounders <- c('age', 'educ', 'black', 'hisp', 'married', 'nodegr')
model_results <- bartCause::bartc(
 response = lalonde[['re78']],
 treatment = lalonde[['treat']],
 confounders = as.matrix(lalonde[, confounders]),
 estimand = 'ate',
 commonSup.rule = 'none'
)
plot_cate_test(model_results, c('age', 'educ'))
}
\author{
George Perrett, Joe Marlo
}
