% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_balance.R
\name{plot_covariance}
\alias{plot_covariance}
\title{Plot the covariance}
\usage{
plot_covariance(.data, treat, confounders)
}
\arguments{
\item{.data}{dataframe}

\item{confounders}{character list of column names denoting the X columns of interest}

\item{treatment}{the column denoted treatment. Must be binary.}
}
\value{
ggplot object
}
\description{
Visualize balance of the covariance of variables between treatment and control groups. Balance plot reflects balance in standardized units.
}
\examples{
data(lalonde)
plot_covariance(lalonde, 'treat', c('re75','re74' , 'age', 'educ')) + labs(title = 'My new title')
}
\author{
George Perrett
}
