% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_moderators.R
\name{plot_moderator_c_bin}
\alias{plot_moderator_c_bin}
\title{Auto-Bin a plot of a continuous moderating variable into a discrete moderating variable}
\usage{
plot_moderator_c_bin(
  .model,
  moderator,
  type = c("density", "histogram", "errorbar"),
  .alpha = 0.7,
  facet = FALSE,
  .ncol = 1,
  .name = "bin"
)
}
\arguments{
\item{.model}{a model produced by `bartCause::bartc()`}

\item{moderator}{the moderator as a vector}

\item{type}{string to specify if you would like to plot a histogram, density or error bar plot}

\item{.alpha}{transparency value [0, 1]}

\item{facet}{TRUE/FALSE. Create panel plots of each moderator level?}

\item{.ncol}{number of columns to use when faceting}

\item{.name}{sting representing the name of the moderating variable}
}
\value{
ggplot object
}
\description{
Use a regression tree to optimally bin a continous variable
}
\examples{
\donttest{
data(lalonde)
confounders <- c('age', 'educ', 'black', 'hisp', 'married', 'nodegr')
model_results <- bartCause::bartc(
 response = lalonde[['re78']],
 treatment = lalonde[['treat']],
 confounders = as.matrix(lalonde[, confounders]),
 estimand = 'ate',
 commonSuprule = 'none'
)
plot_moderator_c_bin(model_results, lalonde$age, .name = 'age')
}
}
\author{
George Perrett
}
