% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_moderators.R
\name{plot_moderator_c_loess}
\alias{plot_moderator_c_loess}
\title{Plot the icates conditional on a continuous moderator}
\usage{
plot_moderator_c_loess(.model, moderator, line_color = "blue")
}
\arguments{
\item{.model}{a model produced by bartCause::bartc()}

\item{moderator}{the moderator as a vector}

\item{line_color}{the color of the loess line}
}
\value{
ggplot object
}
\description{
TODO: description
}
\examples{
data(lalonde)
confounders <- c('age', 'educ', 'black', 'hisp', 'married', 'nodegr')
model_results <- bartCause::bartc(
 response = lalonde[['re78']],
 treatment = lalonde[['treat']],
 confounders = as.matrix(lalonde[, confounders]),
 estimand = 'ate',
 commonSuprule = 'none'
)
plot_moderator_c_loess(model_results, lalonde$age)
}
\author{
George Perrett, Joe Marlo
}
