% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overlap_vars.R
\name{plot_overlap_vars}
\alias{plot_overlap_vars}
\title{Plot the overlap of variables}
\usage{
plot_overlap_vars(
  .data,
  treatment_col,
  confounder_cols,
  plt_type = c("histogram", "density")
)
}
\arguments{
\item{.data}{dataframe}

\item{treatment_col}{name of the treatment column within .data}

\item{confounder_cols}{character list of column names denoting confounders within .data}

\item{plt_type}{the plot type, one of c('histogram', 'density'). Defaults to 'histogram'}
}
\value{
ggplot object
}
\description{
Plot histograms showing the overlap between variables by treatment status.
}
\examples{
data(lalonde, package = 'arm')
plot_overlap_vars(
 .data = lalonde,
 treatment_col = 'treat',
 confounder_cols = c('age', 'educ'),
 plt_type = 'Histogram'
)
}
\seealso{
\code{\link{plot_overlap_pScores}}
}
\author{
George Perrett, Joe Marlo
}
