% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_common_support.R
\name{plot_predicted_common_support}
\alias{plot_predicted_common_support}
\title{Plot a regression tree predicting variables with lack of overlap}
\usage{
plot_predicted_common_support(
  .model,
  max_depth = 3,
  rule = c("both", "sd", "chi")
)
}
\arguments{
\item{.model}{a model produced by `bartCause::bartc()`}

\item{max_depth}{a number indicatin the max depth of the tree. Higher numbers are more prone to overfitting.}

\item{rule}{one of c('both', 'sd', 'chi') denoting which rule to use to identify lack of support}
}
\value{
ggplot object
}
\description{
Identify variables that predict lack of overlap
}
\examples{
\donttest{
data(lalonde)
confounders <- c('age', 'educ', 'black', 'hisp', 'married', 'nodegr')
model_results <- bartCause::bartc(
 response = lalonde[['re78']],
 treatment = lalonde[['treat']],
 confounders = as.matrix(lalonde[, confounders]),
 estimand = 'ate',
 commonSuprule = 'none'
)
plot_predicted_common_support (model_results)
plot_predicted_common_support (model_results, max_depth = 2, rule = 'chi')
}
}
\author{
George Perrett, Joseph Marlo
}
