% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residual.R
\name{plot_residual_density}
\alias{plot_residual_density}
\title{Plot the model residual}
\usage{
plot_residual_density(.model)
}
\arguments{
\item{.model}{a model produced by `bartCause::bartc()`}
}
\value{
ggplot object
}
\description{
Visualize balance of variables between treatment and control groups. Balance plot reflects balance in standardized units.
}
\examples{
\donttest{
data(lalonde)
confounders <- c('age', 'educ', 'black', 'hisp', 'married', 'nodegr')
model_results <- bartCause::bartc(
 response = lalonde[['re78']],
 treatment = lalonde[['treat']],
 confounders = as.matrix(lalonde[, confounders]),
 estimand = 'ate',
 commonSup.rule = 'none'
)
plot_residual_density(model_results)
}
}
\author{
George Perrett & Joseph Marlo
}
