% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variable_importance.R
\name{plot_variable_importance}
\alias{plot_variable_importance}
\title{Variable importance of Bayesian Additive Regression Trees}
\usage{
plot_variable_importance(
  .model,
  confounders,
  plot_or_table = c("both", "plot", "table")
)
}
\arguments{
\item{.model}{a model produced by bartCause::bartc(). Typically store$model_results}

\item{confounders}{a character list of column names which should be considered the confounders. Must match the column names used to original fit .model.}

\item{plot_or_table}{type of output. One of c('both', 'plot', 'table').}
}
\value{
a list containing variable importance plot & ordered table of confounders by scaled importance
}
\description{
Fit single regression tree on bartc() icates to produce variable importance plot & table. TODO: describe what the plot is and how it should be used. Reference?
}
\examples{
data(lalonde, package = 'arm')
confounders <- c('age', 'educ', 'black', 'hisp', 'married', 'nodegr')
model_results <- bartCause::bartc(
 response = lalonde[['re78']],
 treatment = lalonde[['treat']],
 confounders = as.matrix(lalonde[, confounders]),
 estimand = 'ate',
 commonSup.rule = 'none'
)
plot_variable_importance(model_results,  c('age', 'educ'))
}
\author{
George Perrett, Joe Marlo
}
