% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_waterfall.R
\name{plot_waterfall}
\alias{plot_waterfall}
\title{Plot a waterfall}
\usage{
plot_waterfall(
  .model,
  descending = TRUE,
  .order = NULL,
  .color = NULL,
  .alpha = 0.5
)
}
\arguments{
\item{.model}{a model produced by bartCause::bartc()}

\item{descending}{order the icates by value?}

\item{.order}{a vector representing a customer order}

\item{.color}{a vector representing colors}

\item{.alpha}{transparency value [0, 1]}
}
\value{
ggplot object
}
\description{
TODO: description
}
\examples{
data(lalonde)
confounders <- c('age', 'educ', 'black', 'hisp', 'married', 'nodegr')
model_results <- bartCause::bartc(
 response = lalonde[['re78']],
 treatment = lalonde[['treat']],
 confounders = as.matrix(lalonde[, confounders]),
 estimand = 'ate',
 commonSuprule = 'none'
)
plot_waterfall(model_results)
}
\author{
George Perrett
}
