% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overlap_pScores.R
\name{propensity_scores}
\alias{propensity_scores}
\title{Calculate propensity scores using BART}
\usage{
propensity_scores(.data, treatment, response, confounders)
}
\arguments{
\item{.data}{dataframe}

\item{treatment}{character. Name of the treatment column within .data}

\item{response}{character. Name of the response column within .data}

\item{confounders}{character list of column names denoting confounders within .data}
}
\value{
a numeric vector of propensity scores
}
\description{
Calculates propensity scores using Bayesian Additive Regression Trees via bartCause::bartc.
}
\examples{
data(lalonde)
propensity_scores(
  .data = lalonde,
  treatment = 'treat',
  response = 're78',
  confounders = c('age', 'educ')
)
}
\seealso{
\code{\link{plot_overlap_pScores}} \code{\link[rpart.plot]{rpart.plot}}
}
