% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_moderators.R
\name{table_moderator_c_bin}
\alias{table_moderator_c_bin}
\title{Auto-Bin a table of a continuous moderating variable into a discrete moderating variable}
\usage{
table_moderator_c_bin(.model, moderator, .name = "bin")
}
\arguments{
\item{.model}{a model produced by `bartCause::bartc()`}

\item{moderator}{the moderator as a vector}

\item{.name}{sting representing the name of the moderating variable}
}
\value{
a data.frame object
}
\description{
Use a regression tree to optimally bin a continuous variable, this function will print out a table with estimates and 95% ci
}
\examples{
\donttest{
data(lalonde)
confounders <- c('age', 'educ', 'black', 'hisp', 'married', 'nodegr')
model_results <- bartCause::bartc(
 response = lalonde[['re78']],
 treatment = lalonde[['treat']],
 confounders = as.matrix(lalonde[, confounders]),
 estimand = 'ate',
 commonSuprule = 'none'
)
table_moderator_c_bin(model_results, lalonde$age, .name = 'age')
}
}
\author{
George Perrett
}
