% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lalonde}
\alias{lalonde}
\title{Lalonde dataset}
\format{
An object of class \code{data.frame} with 445 rows and 12 columns.
}
\source{
\url{https://CRAN.R-project.org/package=arm}
}
\usage{
lalonde
}
\description{
Lalonde dataset
}
\keyword{datasets}
