% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_balance.R
\name{plot_balance}
\alias{plot_balance}
\title{Plot the balance}
\usage{
plot_balance(
  .data,
  treatment,
  confounders,
  compare = c("means", "variance", "covariance"),
  estimand = c("ATE", "ATT", "ATC")
)
}
\arguments{
\item{.data}{dataframe}

\item{treatment}{the column denoted treatment. Must be binary.}

\item{confounders}{character list of column names denoting the X columns of interest}

\item{compare}{character of either means or variance denotes what to compare balance on}

\item{estimand}{character of either ATE, ATT or ATC the causal estimand you are making inferences about}
}
\value{
ggplot object
}
\description{
Visualize balance of variables between treatment and control groups. Balance plot reflects balance in standardized units.
}
\examples{
data(lalonde)
plot_balance(lalonde, 'treat', c('re78', 'age', 'educ'), compare = 'means', estimand = 'ATE') + labs(title = 'My new title')
}
\author{
Joseph Marlo & George Perrett
}
