% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_balance.R
\name{plot_balance}
\alias{plot_balance}
\title{Plot the balance}
\usage{
plot_balance(.data, z_col, x_cols)
}
\arguments{
\item{z_col}{the column denoted treatment. Must be binary.}

\item{x_cols}{character list of column names denoting the X columns of interest.}

\item{X}{dataframe}
}
\value{
ggplot object
}
\description{
Visualize balance of variables between treatment and control groups."
}
\author{
Joe Marlo
}
