% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ATE.R
\name{plot_CATE}
\alias{plot_CATE}
\title{Plot histogram or density of individual treatment effects}
\usage{
plot_CATE(
  .model,
  type = c("histogram", "density"),
  ci_80 = FALSE,
  ci_95 = FALSE,
  reference = NULL,
  .mean = FALSE,
  .median = FALSE
)
}
\arguments{
\item{.model}{a model produced by bartCause::bartc()}

\item{type}{histogram or density}

\item{ci_80}{TRUE/FALSE. Show the 80\% credible interval?}

\item{ci_95}{TRUE/FALSE. Show the 95\% credible interval?}

\item{reference}{numeric. Show a vertical reference line at this value}

\item{.mean}{TRUE/FALSE. Show the mean reference line}

\item{.median}{TRUE/FALSE. Show the median reference line}
}
\value{
ggplot object
}
\description{
TODO: description
}
\examples{
data(lalonde)
confounders <- c('age', 'educ', 'black', 'hisp', 'married', 'nodegr')
model_results <- bartCause::bartc(
 response = lalonde[['re78']],
 treatment = lalonde[['treat']],
 confounders = as.matrix(lalonde[, confounders]),
 estimand = 'ate',
 commonSup.rule = 'none'
)
plot_CATE(model_results)
}
\author{
George Perrett, Joe Marlo
}
