% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostic_common_support.R
\name{plot_diagnostic_common_support}
\alias{plot_diagnostic_common_support}
\title{Plot common support based on the standard deviation rule, chi squared rule or both}
\usage{
plot_diagnostic_common_support(
  .model,
  rule = c("none", "sd", "chi"),
  plot_theme = ggplot2::theme_get
)
}
\arguments{
\item{.model}{a model produced by bartCause::bartc()}

\item{rule}{one of c('none', 'sd', 'chi') denoting which rule to use to identify lack of support}

\item{plot_theme}{a ggplot2 theme function. Defaults to the current global theme}
}
\value{
ggplot object
}
\description{
Returns a ggplot common support plot. TODO: TODO: describe what the plot is and how it should be used. reference?
}
\examples{
data(lalonde)
confounders <- c('age', 'educ', 'black', 'hisp', 'married', 'nodegr')
model_results <- bartCause::bartc(
 response = lalonde[['re78']],
 treatment = lalonde[['treat']],
 confounders = as.matrix(lalonde[, confounders]),
 estimand = 'ate',
 commonSuprule = 'none'
)
plot_diagnostic_common_support(model_results, plot_theme = ggplot2::theme_minimal)
}
\author{
George Perrett, Joe Marlo
}
