% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ATE.R
\name{plot_ICATE}
\alias{plot_ICATE}
\title{Plot icates}
\usage{
plot_ICATE(.model, group.by = NULL, nbins = 30)
}
\arguments{
\item{.model}{a model produced by bartCause::bartc()}

\item{group.by}{a grouping variable as a vector}

\item{nbins}{number of bins #TODO describe}
}
\value{
ggplot object
}
\description{
TODO: description
}
\examples{
data(lalonde)
confounders <- c('age', 'educ', 'black', 'hisp', 'married', 'nodegr')
model_results <- bartCause::bartc(
 response = lalonde[['re78']],
 treatment = lalonde[['treat']],
 confounders = as.matrix(lalonde[, confounders]),
 estimand = 'ate',
 commonSup.rule = 'none'
)
plot_ICATE(model_results, lalonde$married)
}
\author{
George Perrett
}
