% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ATE.R
\name{plot_ICATE}
\alias{plot_ICATE}
\title{Plot icates}
\usage{
plot_ICATE(.model, .group_by = NULL, n_bins = 30, .alpha = 0.7)
}
\arguments{
\item{.model}{a model produced by bartCause::bartc()}

\item{.group_by}{a grouping variable as a vector}

\item{n_bins}{number of bins. Defaults to 30}

\item{.alpha}{transparency of histograms}
}
\value{
ggplot object
}
\description{
TODO: description
}
\examples{
\dontrun{
data(lalonde)
confounders <- c('age', 'educ', 'black', 'hisp', 'married', 'nodegr')
model_results <- bartCause::bartc(
 response = lalonde[['re78']],
 treatment = lalonde[['treat']],
 confounders = as.matrix(lalonde[, confounders]),
 estimand = 'ate',
 commonSup.rule = 'none'
)
plot_ICATE(model_results, lalonde$married)
}
}
\author{
George Perrett
}
