% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_moderators.R
\name{plot_moderator_c_pd}
\alias{plot_moderator_c_pd}
\title{Plot the icates conditional on a continuous moderator}
\usage{
plot_moderator_c_pd(
  .model,
  moderator,
  n_bins = 15,
  legend = c("none", "right", "top", "bottom")
)
}
\arguments{
\item{.model}{a model produced by bartCause::bartc()}

\item{moderator}{the moderator as a vector}

\item{n_bins}{number of bins}

\item{legend}{legend position. One of c('none', 'right', 'top', 'bottom')}
}
\value{
ggplot object
}
\description{
TODO: description
}
\examples{
data(lalonde)
confounders <- c('age', 'educ', 'black', 'hisp', 'married', 'nodegr')
model_results <- bartCause::bartc(
 response = lalonde[['re78']],
 treatment = lalonde[['treat']],
 confounders = as.matrix(lalonde[, confounders]),
 estimand = 'ate',
 commonSuprule = 'none',
 keepTrees = TRUE
)
plot_moderator_c_pd(model_results, lalonde$age)
}
\author{
George Perrett, Joe Marlo
}
