% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_moderator_search.R
\name{plot_moderator_search}
\alias{plot_moderator_search}
\title{Single Regression Tree for exploratory heterogenious effects}
\usage{
plot_moderator_search(
  .model,
  depth = 2,
  type = c(2, 0, 1, 3, 4, 5),
  extra = list(1, "all", 0)
)
}
\arguments{
\item{.model}{a model produced by bartCause::bartc(). Typically store$model_results}

\item{confounders}{matrix of confounders}
}
\value{
a ggplot object of a single regression tree
}
\description{
Fit single regression tree on bartc() icates to produce variable importance plot & table.
}
\author{
George Perrett, Joe Marlo
}
