% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_moderators.R
\name{plot_moderator_search}
\alias{plot_moderator_search}
\title{Plot a single regression tree for exploratory heterogeneous effects}
\usage{
plot_moderator_search(
  .model,
  depth = 2,
  type = c(2, 0, 1, 3, 4, 5),
  extra = list(1, "all", 0)
)
}
\arguments{
\item{.model}{a model produced by bartCause::bartc(). Typically store$model_results}

\item{depth}{number of node levels within the tree. 2 is recommended}

\item{type}{type of plot to draw. See \link[rpart.plot]{rpart.plot} for details}

\item{extra}{extra information to display. See \link[rpart.plot]{rpart.plot} for details}
}
\value{
rpart plot
}
\description{
Fit single regression tree on bartc() icates to produce variable importance plot & table.
}
\author{
George Perrett, Joe Marlo
}
