% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overlap_pScores.R
\name{plot_overlap_pScores}
\alias{plot_overlap_pScores}
\title{Plot the overlap via propensity score method}
\usage{
plot_overlap_pScores(
  .data,
  treatment,
  response,
  confounders,
  plot_type = c("histogram", "density")
)
}
\arguments{
\item{.data}{dataframe}

\item{treatment}{name of the treatment column within .data}

\item{response}{name of the response column within .data}

\item{confounders}{character list of column names denoting confounders within .data}

\item{plot_type}{the plot type, one of c('Histogram', 'Density')}
}
\value{
ggplot object
}
\description{
Plot histograms showing the overlap between propensity scores by treatment status. TODO: reference?
}
\examples{
data(lalonde, package = 'arm')
plot_overlap_pScores(
 .data = lalonde,
 treatment = 'treat',
 response = 're78',
 confounders = c('age', 'educ'),
 plot_type = 'histogram'
)
}
\seealso{
\code{\link{plot_overlap_vars}}
}
\author{
George Perrett, Joe Marlo
}
